Fix sync issue where server gets input from before last sync
Server sends a sync message at time 100
Receives an input for player at time 99

In this case, server will roll back to 99, apply the input, and
roll forward, so the server is good. But the clients will be stuck
on the stale sync message till the next sync (for up to a second!).

Now, whenever server receives an input that's older than the last
sync message, it immediately sends a new sync message to "correct"
the last sync.