Improve rocket firing latency
Now the client will also create and simulate a rocket fire, instead
of waiting for the server to send the sync. This is necessary
due to the previous change (bc26e096a2d493b60686c74803a3c1ae3f919b49)
which made the client drop sync messages when it perceives the sync
message may be incorrect because of unapplied inputs. But this led to
sync messages with rockets being dropped sometimes, leading to a
noticeable lag.

Now the client will also immediately create the rocket. But this is also
made possible and easy due to that same change -- we will now drop
any sync message that doesn't incorporate inputs we know about, which
means we will not apply sync messages that don't include this rocket,
so there's no risk of a sync that suddenly removes the rocket that we
just fired because the server was a little behind. Instead, now we will
only accept sync messages that will include the rocket (that is, we only
accept sync messages where the server says it has seen the rocket-firing
input).  Should the rocket be ILLEGALLY fired (like there's actually
not enough ammo for it), the server will process the fire event and refuse
to add a rocket, and the client will then dutifully receive the sync
message and remove the rocket from its world state.